Correccion PR feature/OT185-31
