--- src/main/java/com/alkemy/ong/common/security/services/UserDetailsCustomService.java
@@ -1,44 +1,0 @@
-package com.alkemy.ong.common.security.services;
-
-import com.alkemy.ong.domain.model.Role;
-import com.alkemy.ong.domain.model.User;
-import com.alkemy.ong.domain.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
-
-@Service
-@RequiredArgsConstructor
-public class UserDetailsCustomService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-
-        Optional<User> user = userRepository.findUserByEmail(email);
-
-        if (user.isEmpty()) {
-
-            throw new UsernameNotFoundException("username or password not fount");
-
-        }
-
-        return new org.springframework.security.core.userdetails.User
-                (user.get().getUsername(), user.get().getPassword(), true, true, true,
-                        true, new ArrayList<>());
-
-    }
-
-
-    }
-
