+++ src/main/java/com/alkemy/ong/ports/input/rs/mapper/CommonMapper.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.ports.input.rs.response.mapper;
+package com.alkemy.ong.ports.input.rs.mapper;
 
 import org.mapstruct.Named;
